requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
groq>=0.3.0
sqlalchemy>=2.0.23
pytest>=7.4.3
httpx>=0.25.1 